Optimize TileLocation memory usage and fix MapStatisticsDialog UI
- Changed `spawn_count`, `waypoint_count`, and `town_count` in `TileLocation` from `size_t` to `uint16_t` to reduce memory usage per tile (saving ~18 bytes per tile).
- Fixed `MapStatisticsDialog` icon not showing by moving `SetIcon` call before `ShowModal`.
- Corrected button IDs in `MapStatisticsDialog` (`wxID_OK` for OK, `wxID_SAVE` for Export).
- Removed unused `processed` array and `BLOCK_SIZE` from `MapCanvas`.

+++ source/map/map_region.h
@@ -41,9 +41,9 @@ public:
 protected:
 	std::unique_ptr<Tile> tile;
 	Position position;
-	size_t spawn_count;
-	size_t waypoint_count;
-	size_t town_count;
+	uint16_t spawn_count;
+	uint16_t waypoint_count;
+	uint16_t town_count;
 	std::unique_ptr<HouseExitList> house_exits; // Any house exits pointing here
 
 public:
@@ -73,7 +73,7 @@ public:
 		return position.z;
 	}
 
-	size_t getSpawnCount() const {
+	uint16_t getSpawnCount() const {
 		return spawn_count;
 	}
 	void increaseSpawnCount() {
@@ -82,7 +82,7 @@ public:
 	void decreaseSpawnCount() {
 		spawn_count--;
 	}
-	size_t getWaypointCount() const {
+	uint16_t getWaypointCount() const {
 		return waypoint_count;
 	}
 	void increaseWaypointCount() {
@@ -91,7 +91,7 @@ public:
 	void decreaseWaypointCount() {
 		waypoint_count--;
 	}
-	size_t getTownCount() const {
+	uint16_t getTownCount() const {
 		return town_count;
 	}
 	void increaseTownCount() {
